+++ src/main/java/com/project/webshopproject/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.project.webshopproject.repository;
 
-import com.project.webshopproject.entity.User;
+import com.project.webshopproject.entity.user.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
