+++ src/main/java/com/project/webshopproject/security/UserDetailsServiceImpl.java
@@ -1,6 +1,6 @@
 package com.project.webshopproject.security;
 
-import com.project.webshopproject.entity.User;
+import com.project.webshopproject.entity.user.User;
 import com.project.webshopproject.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
